Extract duplicated code into abstract class
